refactor: re-organize command implementations in multiple files
